+++ src/main/kotlin/fr/polyflix/search/infrastructure/messaging/kafka/consumer/KafkaVideoConsumer.kt
@@ -16,6 +16,12 @@ class KafkaVideoConsumer(private val videoRepository: VideoRepository) {
     fun onMessage(event: VideoEvent, ack: Acknowledgment) {
         logger.info("Received new event: $event")
 
+        // Ignore draft elements and non public elements
+        if (event.payload.draft == true || event.payload.visibility != "public") {
+            logger.info("The video ${event.payload.id} isn't public or it is in draft state. Ignoring it.")
+            return ack.acknowledge()
+        }
+
         when (event.trigger) {
             Trigger.CREATE, Trigger.UPDATE -> videoRepository.createOne(event.payload)
             Trigger.DELETE -> videoRepository.deleteOne(event.payload.id!!)
